+++ app/domain/annotaitons/annotations_handler.rb
@@ -4,8 +4,12 @@ module AnnotationsHandler
     resource.annotations_dataset.import(%i[resource_id name value policy_id], records)
   end
 
-  def annotation_value_by_name(secret, name)
-    annotations = secret.annotations
+  def delete_resource_annotations(resource)
+    Annotation.where(resource_id: resource.id).delete
+  end
+
+  def annotation_value_by_name(resource, name)
+    annotations = resource.annotations
     target_annotation = annotations.find { |annotation| annotation.name == name }
     if target_annotation
       target_annotation.value