+++ spec/app/domain/secrets/secret_types/base_secret_type_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-describe "Base secret input validation" do
+describe "Base secret create input validation" do
   let(:secret) do
     Secrets::SecretTypes::SecretBaseType.new
   end
@@ -12,70 +12,70 @@ describe "Base secret input validation" do
   context "when creating secret with empty name" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "", branch: "data")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(Errors::Conjur::ParameterMissing)
     end
   end
   context "when creating secret with no name" do
     it "input validation fails" do
       params = ActionController::Parameters.new(branch: "data")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(Errors::Conjur::ParameterMissing)
     end
   end
   context "when creating secret with name not string" do
     it "input validation fails" do
       params = ActionController::Parameters.new(branch: "data", name: 5)
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(Errors::Conjur::ParameterTypeInvalid)
     end
   end
   context "when creating secret with unsupported symbols in its name" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "se#cret/not_valid", branch: "data")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(ApplicationController::BadRequestWithBody)
     end
   end
   context "when creating secret with too long name" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "secretstoolongggggggggggggggggggggggggggggggggggggggggggggggg", branch: "data")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(ApplicationController::BadRequestWithBody)
     end
   end
   context "when creating secret with all supported symbols in its name" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "seCret0_5Name", branch: "data")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to_not raise_error
     end
   end
   context "when creating secret with only numbers in its name" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "12345", branch: "data")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to_not raise_error
     end
   end
   context "when creating secret with empty branch" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "secret1", branch: "")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(Errors::Conjur::ParameterMissing)
     end
   end
   context "when creating secret with no branch" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "secret1")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(Errors::Conjur::ParameterMissing)
     end
   end
   context "when creating secret with branch not string" do
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "secret1", branch: 5)
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(Errors::Conjur::ParameterTypeInvalid)
     end
   end
@@ -86,7 +86,7 @@ describe "Base secret input validation" do
     end
     it "input validation fails" do
       params = ActionController::Parameters.new(name: "secret1", branch: "luba")
-      expect { secret.input_validation(params)
+      expect { secret.create_input_validation(params)
       }.to raise_error(Exceptions::RecordNotFound)
     end
   end